pytest app fixture now cleans up the db for each test

+++ tests/conftest.py
@@ -25,7 +25,9 @@ def app():
     db_path = os.path.join(
         pathlib.Path(__file__).resolve().parent.parent, "picure/DB/picure.sqlite"
     )
-    os.remove(db_path)
+    if os.path.isfile(db_path):
+        os.remove(db_path)
+
     return create_app(
         {
             "TESTING": True,
